+++ polaris/dataset/__init__.py
@@ -1,17 +1,19 @@
 from polaris.dataset._column import ColumnAnnotation, KnownContentType, Modality
 from polaris.dataset._dataset import DatasetV1, DatasetV1 as Dataset
+from polaris.dataset._dataset_v2 import DatasetV2
 from polaris.dataset._factory import DatasetFactory, create_dataset_from_file, create_dataset_from_files
 from polaris.dataset._subset import Subset
 
+
 __all__ = [
-    "ColumnAnnotation",
-    "Dataset",
-    "Subset",
-    "Modality",
-    "KnownContentType",
-    "DatasetFactory",
     "create_dataset_from_file",
     "create_dataset_from_files",
-    "DatasetV1",
+    "ColumnAnnotation",
     "Dataset",
+    "DatasetFactory",
+    "DatasetV1",
+    "DatasetV2",
+    "KnownContentType",
+    "Modality",
+    "Subset",
 ]
